requirements.txt
cycler==0.11.0
fonttools==4.34.4
kiwisolver==1.4.3
matplotlib==3.5.2
numpy==1.23.0
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
plotly==5.9.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2022.1
tenacity==8.0.1
